add user id to app

+++ server/repository/spotify.repository.ts
@@ -1,17 +1,21 @@
 import type { SpotifyPlaylistI } from "@model/spotify/playlist";
 import type { Artist, SavedArtistI } from "@model/spotify/search";
+import type { User } from "@model/spotify/user";
 import { type Collection, type Db, MongoClient } from "mongodb";
 
-export function SpotifyRepository(mongoUri: string) {
-	let client: MongoClient;
-	let db: Db;
-	let artistDb: Collection;
-	let playlistDb: Collection;
+let client: MongoClient;
+let db: Db;
+let userDb: Collection;
+let artistDb: Collection;
+let playlistDb: Collection;
 
+export function SpotifyRepository(mongoUri: string) {
 	async function connect(): Promise<void> {
 		client = new MongoClient(mongoUri);
 		await client.connect();
+
 		db = client.db("spotify");
+		userDb = db.collection("users");
 		artistDb = db.collection("saved_artists");
 		playlistDb = db.collection("playlist");
 		console.log("Connected to MongoDB.");
@@ -24,11 +28,23 @@ export function SpotifyRepository(mongoUri: string) {
 		}
 	}
 
-	async function fetchAndSaveArtists(artists: Array<Artist>) {
+	async function saveUser(userData: User) {
+		return await userDb.updateOne(
+			{ userId: userData.userId },
+			{ $set: userData },
+			{ upsert: true },
+		);
+	}
+
+	async function getUser(userId: string) {
+		return await userDb.findOne<User>({ userId });
+	}
+
+	async function fetchAndSaveArtists(userId: string, artists: Array<Artist>) {
 		const bulkOperations = artists.map((artist) => ({
 			updateOne: {
-				filter: { id: artist.id },
-				update: { $set: artist },
+				filter: { userId, id: artist.id },
+				update: { $set: { ...artist, userId } },
 				upsert: true,
 			},
 		}));
@@ -36,30 +52,38 @@ export function SpotifyRepository(mongoUri: string) {
 		return await artistDb.bulkWrite(bulkOperations);
 	}
 
-	async function resetArtists(artists: Array<SavedArtistI>) {
-		await artistDb.deleteMany({});
-		return await artistDb.insertMany(artists);
+	async function resetArtists(userId: string, artists: Array<SavedArtistI>) {
+		await artistDb.deleteMany({ userId });
+		return await artistDb.insertMany(
+			artists.map((artist) => ({ ...artist, userId })),
+		);
 	}
 
-	async function removeSavedArtist(id: string) {
-		return await artistDb.deleteOne({ id });
+	async function removeSavedArtist(userId: string, id: string) {
+		const result = await artistDb.deleteOne({ userId, id });
+		if (result.deletedCount === 0) {
+			throw new Error(`Artist not found for user ${userId}`);
+		}
+		return result;
 	}
 
-	async function getAllArtists() {
-		return await artistDb.find({}).toArray();
+	async function getAllArtists(userId: string) {
+		return await artistDb.find<Artist>({ userId }).toArray();
 	}
 
-	async function getPlaylist() {
-		return await playlistDb.findOne<SpotifyPlaylistI>({});
+	async function getPlaylist(userId: string) {
+		return await playlistDb.findOne<SpotifyPlaylistI>({ userId });
 	}
 
-	async function createPlaylist(playlist: SpotifyPlaylistI) {
-		return await playlistDb.insertOne(playlist);
+	async function createPlaylist(userId: string, playlist: SpotifyPlaylistI) {
+		return await playlistDb.insertOne({ ...playlist, userId });
 	}
 
 	return {
 		connect,
 		disconnect,
+		saveUser,
+		getUser,
 		fetchAndSaveArtists,
 		resetArtists,
 		removeSavedArtist,
